requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-jinja2==1.6
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
black==25.12.0
click==8.3.1
decorator==5.2.1
executing==2.2.1
frozenlist==1.8.0
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-json-logger==4.0.0
pytokens==0.3.0
PyYAML==6.0.3
rich==14.2.0
stack-data==0.6.3
textual==6.8.0
textual-dev==1.8.0
textual-serve==1.1.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
wcwidth==0.2.14
yarl==1.22.0
